+++ conf/readinglist.ini
@@ -3,6 +3,7 @@ use = egg:readinglist
 pyramid.debug_notfound = true
 readinglist.backend = readinglist.backend.memory
 readinglist.http_scheme = http
+readinglist.retry_after = 30
 readinglist.eos =
 # readinglist.basic_auth_backdoor = bob:secret
 # readinglist.backoff = 10
