+++ readinglist/tests/test_views_errors.py
@@ -43,11 +43,12 @@ class ErrorViewTest(BaseWebTest, unittest.TestCase):
             "This user cannot access this resource.")
 
     def test_500_is_valid_formatted_error(self):
-        with mock.patch('traceback.print_exc', return_value=True):
+        with mock.patch('traceback.format_exc', return_value="") as mock_err:
             with mock.patch('readinglist.views.article.Article.collection_get',
                             side_effect=ValueError):
                 response = self.app.get('/articles',
                                         headers=self.headers, status=500)
+        mock_err.assert_called_once_with()
         self.assertFormattedError(
             response, 500, ERRORS.UNDEFINED, "Internal Server Error",
             "A programmatic error occured, developers have been informed.")
@@ -58,3 +59,4 @@ class ErrorViewTest(BaseWebTest, unittest.TestCase):
         self.assertFormattedError(
             response, 503, ERRORS.BACKEND, "Service unavailable",
             "Service unavailable due to high load, please retry later.")
+        self.assertIn("Retry-After", response.headers)
